WIP: implemented basic functionality to DetailsScreen.kt and DetailsViewModel.kt to test getting movie details. Fixed mistake in AbstractionsModule.kt where incorrect variable was passed.

+++ app/src/main/java/com/myapplications/mywatchlist/ui/MainActivity.kt
@@ -148,7 +148,7 @@ class MainActivity : ComponentActivity() {
                                 navArgument("titleType") {type = NavType.StringType}
                             )
                         ) { backStackEntry ->
-                            val titleId = backStackEntry.arguments?.getInt("titleId")
+                            val titleId = backStackEntry.arguments?.getLong("titleId")
                             val titleType = backStackEntry.arguments?.getString("titleType")
                             if (titleId != null && titleType != null) {
                                 DetailsScreen(titleId = titleId, titleType = titleType)
